MTDSA-25645 CIS Deductions API - EMA Implementation
