+++ app/v1/connectors/AmendConnector.scala
@@ -27,11 +27,17 @@ import v1.models.request.amend.AmendRequestData
 import javax.inject.Inject
 import scala.concurrent.{ExecutionContext, Future}
 
-class AmendConnector @Inject() (val http: HttpClient, val appConfig: AppConfig)(implicit featureSwitches: CisDeductionsApiFeatureSwitches)
-    extends BaseDownstreamConnector {
+class AmendConnector @Inject() (val http: HttpClient, val appConfig: AppConfig)(implicit
+    featureSwitches: CisDeductionsApiFeatureSwitches
+) extends BaseDownstreamConnector {
 
   def amendDeduction(
-      request: AmendRequestData)(implicit hc: HeaderCarrier, ec: ExecutionContext, correlationId: String): Future[DownstreamOutcome[Unit]] = {
+      request: AmendRequestData
+  )(implicit
+      hc: HeaderCarrier,
+      ec: ExecutionContext,
+      correlationId: String
+  ): Future[DownstreamOutcome[Unit]] = {
 
     import request._
